created individual test files for each class
